Unbound: rooms > 127, dyna actors > 50, world extent > +/-32767
Three of the "known remaining limits" from unbound-docs/README.md.

Rooms (counts.md)
- Room numbers s8 -> s16 everywhere (Room.num, Actor.room, transition and
  entrance entries + C++ mirrors, save-side fields, narrow locals, port-layer
  readers). Binary loaders read the room byte signed so vanilla 0xFF is still -1.
- The real cap was 32: clear/temp-clear flags were 1 << room on a u32. Rooms
  >= 32 use SceneFlagsExt_* (growable per-scene bitset, saved by scene name in
  the "unbound" section). Minimap bits are guarded to rooms < 32.
- WaterBox.room is unpacked at load; collision.json accepts an explicit "room"
  (-1 = all) that lifts the 6-bit / 63-room cap.
- Transition actors: Actor.transitionIndex + TRANSITION_ACTOR_INDEX() replace
  the 6-bit index packed into params (cap was 64); numActors is u16.

Dyna collision (collision.md)
- bgActors/bgActorFlags are heap tables (64, doubling); BG_ACTOR_MAX removed,
  BGCHECK_SCENE is a fixed sentinel, BGACTOR_INVALID names the spawn-failure
  value in the ~60 overlay tests.
- Actor.floorBgId/wallBgId u8 -> s32; Camera bg ids s16 -> s32.
- Dyna poly/vtx lists grow in DynaPoly_Setup; superseded buffers are retired
  until BgCheck_Free so CollisionPoly* held by actors stay readable.
- Fix: the "transform unchanged" branch of DynaPoly_ExpandSRT passed an s16
  to the widened s32* DynaSSNodeList_SetSSListHead.

World extent (extent.md)
- Float Mtx via the libultraship fork (GBI_FLOAT_MTX); guMtxF2L/L2F are
  copies, gMtxClear is a float identity, the hand-packed writers in
  sys_matrix.c and MtxConv_F2L build MtxF.
- rooms/<n>.json "origin": Room.origin, applied by Room_OriginMtx in z_room.c.
- f32 positions: ActorEntry/TransitionActorEntry/PolygonDlist2 pos, Path
  points, LightPoint, HorseData.pos, collision vertices/dist/bounds, WaterBox;
  collision.bin v2 (unbound/collision/2), v1 still loads; converter emits v2.
- Sphere16/Cylinder16 and ColliderQuad mid-points are f32 fields, so the
  Math3D_/Collider_ API and ~200 overlays compile unchanged.
- BGCHECK_XYZ_ABSMAX 2^20, BGCHECK_Y_MIN -2^31, effect cull follows.

Also
- `soh --export-unbound` no longer hangs: RunExtract treated argv as ROMs.
- Docs: extent.md, prelude-handoff.md, plan-extent-rooms-dyna.md; README,
  counts.md, collision.md, scene-format.md updated.

+++ soh/soh/Enhancements/Warping.cpp
@@ -20,7 +20,7 @@ void Select_LoadGame(SelectContext* selectContext, s32 entranceIndex);
 
 typedef struct WarpPoint {
     s32 entranceId;
-    s8 roomNum;
+    s16 roomNum; // SOH [Unbound] s8 -> s16
     Vec3f pos;
     s16 rotY;
     bool bootToPoint;
